move iterator to utils
