+++ package.nls.json
@@ -41,7 +41,6 @@
 "mssql.azure.enablePublicCloudDescription":"Should Azure public cloud integration be enabled",
 "mssql.azure.enableUsGovCloudDescription":"Should US Government Azure cloud (Fairfax) integration be enabled",
 "mssql.azure.enableChinaCloudDescription":"Should Azure China cloud integration be enabled",
-
 "mssql.connection.server":"[Required] Specify the server name to connect to. Use 'hostname instance' or '<server>.database.windows.net' for Azure SQL Database.",
 "mssql.connection.database":"[Optional] Specify the database name to connect to. If database is not specified, the default user database setting is used, typically 'master'.",
 "mssql.connection.user":"[Optional] Specify the user name for SQL Server authentication. If user name is not specified, when you connect, you will be asked again.",
