+++ src/utils/iterator.ts
